backend/requirements.txt
streamlit
pandas
scikit-learn
nltk
fastapi
uvicorn
requests
faiss-cpu>=1.7.4
urllib3<2
tensorflow
torch
torchtext
passlib[bcrypt]
bcrypt<4.0.0
PyJWT
sqlalchemy
python-multipart
vaderSentiment
numpy
psycopg2-binary
